+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The `pycldf` package adheres to [Semantic Versioning](http://semver.org/spec/v2.
 ## [Unreleased]
 
 - Support reading and writing sources from/to zipped BibTeX files.
+- Load `Dataset.sources` lazily. (See https://github.com/cldf/pycldf/issues/162)
 
 
 ## [1.33.0] - 2022-11-24